requirements.txt
cmake==3.17.2
dlib==19.18.0
opencv-python==4.6.0.66
face_recognition==1.3.0
easydict==1.10
numpy==1.23.5
tqdm==4.64.1
torchvision==0.14.1
torch==1.13.1
Pillow==9.3.0
tensorboardX==2.5.1
